iTerm shell integration is now supported
iTerm's shell integration provides a way for terminal sessions to make
information available to the iTerm application. I'm adding this because
I'd like to be able to have the short version (minus any trailing
".gatech.edu") host name available for use in an iTerm badge.

In support of this, I've also put iTerm's own
.iterm2_shell_integration.zsh script into this project. This might not
be the technically proper way to do this, but I'm really the only one
using this project. But I WILL have to update that script in this
project when the iTerm folks distribute a new version of the file.