Even BETTER command history
When the home directory is mounted over the network, we include the
local hostname in the name of the history file. This way, all sessions
on THAT host share history, but no host pollutes the history of another.

Also, the histfcntllock option is now set ONLY when the home directory
is mounted over the network.