Git branch support
If the current directory is in a Git repository, the window title now
includes the name of the current branch of that repository.

This also includes a new git_branch Zshell function, which can be called
at any time (from an interactive shell). It returns either the name of
the current Git branch or nothing, which indicates that PWD is not under
a Git repository.