Tweaked log function output
I brightened the log colors in general and set debug to use light gray
on blue, which sets it off nicely from the other log levels.

I've also set up all the log functions to color the timestamp prefixing
the log output. This is also to draw attention to it and to better
distinguish it from normal script output.