Bug fix for install & add .bash_profile
Since I've added .bash_profile as a symlink to .zshenv and .bashrc as a
symlink to .zshrc, I had to fix a bug that munged up the installion of
symlinks that pointed to symlinks. The installer's much happier now, but
I still need to put bash through its paces under the resulting regime.