Improve RHEL and AMI detection in .zshenv
On both Red Hat Enterprise Linux and Amazon Linux AMI platforms, the
/etc/zlogout file exists ONLY and SADISTICALLY to clear the screen,
presumably for the sake of security, but effectively preventing the user
from seeing the end of the closed session. I hate this with the burning
passion of a billion hells, so I like to "setopt noglobalrcs" when using
Zshell on these platforms.

The problem was that on one of my machine, shells are symlinked so they
look like the logged-in username (not my idea). So now I'm using my
"realpath" shell function to resolve SHELL's value before looking to see
if it's Zshell.