Added fw and fwe aliases for making fireworks
These two aliases run "it2attention fireworks", and fwe exits the shell
once the fireworks finish. I'm thinking I'll add ";fwe" at the end of
long-running root commands so that I'll notice when they're done, and if
I'm not around to see it, the shell will exit (to avoid hanging because
of inactivity).