Add temperature conversion functions
Usage: celsius TEMP
       fahrenheit TEMP

Convert between Celsius and Fahrenheit degrees. The celsius function
converts from Fahrenheit to Celsius degrees, and the fahrenheit function
converts from Celsius to Fahrenheit degrees.

The TEMP argument must be a base-10, possibly floating point number. It
may be expressed in either standard or scientific notation, and values
on the interval (-1,1) needn't have a 0 to the left of the decimal
point.